initials: silence "populate_sysroot manifest not found" warnings
A number of recipes in meta-java use autotools.bbclass for building; internally
it uses the sstate-control populate_sysroot manifests to locate and copy
aslocals for configure dependencies as part of the do_configure step.
The manifest path differs depending on if it's a native package or not so
autotools.bbclass looks at the package name to determine if its native or not
(it's native if it ends with "-native").

The warnings are emitted because a few (native) recipes in meta-java (such as
cacao-initial, classpath-initial and more) are incorrectly being classified as
non-native by autotools.bbclass, which makes it look for the manifest in the
wrong place and finally gives up with the warning:

WARNING: /home/oe/tmp-glibc/sstate-control/manifest-mymachine-libecj-bootstrap.populate_sysroot not found

This happens for cacao-initial, classpath-initial, ecj-initial, libecj-boostrap,
jamvm-initial and jikes-initial since they are all native and rely on autotools.

This patch renames the recipes so they end with -native, e.g ecj-initial-native.
It also updates the recipes so they DEPEND on the corresponding new name. This
helps autotools.bbclass to properly classify the recipes are native and silence
the warnings.

Signed-off-by: Magnus Olsson <magnus@minimum.se>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>